Conflicts:
	Prelo/Classes/Model/API/PreloEndpoints.swift
	Prelo/Classes/Model/CoreData/CDVersion.swift

+++ Prelo/AppDelegate.swift
@@ -93,28 +93,61 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                         
                         // Jika versi metadata baru, load dan save kembali di coredata
                         let ver : CDVersion? = CDVersion.getOne()
-                        if (ver?.metadataVersion == data["metadata_version"].string) {
-                            println("Same metadata version")
-                            
-                            // cek validasi data
-                            let arr = CDProvince.getProvincePickerItems()
-                            if (arr.count == 0)
-                            {
-                                self.updateMetadata()
-                            }
+                        var isUpdate : Bool = false
+                        var isUpdateVers : [String] = []
+                        if (ver?.brandsVersion == data["metadata_versions"]["brands"].number! && CDBrand.getBrandCount() > 0) {
+                            isUpdateVers.append("0")
                         } else {
-                            self.updateMetadata()
+                            isUpdateVers.append("1")
+                            isUpdate = true
                         }
-                        CDVersion.saveVersion(data)
+                        if (ver?.categoriesVersion == data["metadata_versions"]["categories"].number! && CDCategory.getCategoryCount() > 0) {
+                            isUpdateVers.append("0")
+                        } else {
+                            isUpdateVers.append("1")
+                            isUpdate = true
+                        }
+                        if (ver?.categorySizesVersion == data["metadata_versions"]["category_sizes"].number! && CDCategorySize.getCategorySizeCount() > 0) {
+                            isUpdateVers.append("0")
+                        } else {
+                            isUpdateVers.append("1")
+                            isUpdate = true
+                        }
+                        if (ver?.shippingsVersion == data["metadata_versions"]["shippings"].number! && CDShipping.getShippingCount() > 0) {
+                            isUpdateVers.append("0")
+                        } else {
+                            isUpdateVers.append("1")
+                            isUpdate = true
+                        }
+                        if (ver?.productConditionsVersion == data["metadata_versions"]["product_conditions"].number! && CDProductCondition.getProductConditionCount() > 0) {
+                            isUpdateVers.append("0")
+                        } else {
+                            isUpdateVers.append("1")
+                            isUpdate = true
+                        }
+                        if (ver?.provincesRegionsVersion == data["metadata_versions"]["provinces_regions"].number! && CDProvince.getProvinceCount() > 0) {
+                            isUpdateVers.append("0")
+                        } else {
+                            isUpdateVers.append("1")
+                            isUpdate = true
+                        }
+                        
+                        // Update jika ada version yg berbeda
+                        if (isUpdate) {
+                            self.updateMetadata(isUpdateVers[0], updateCategories: isUpdateVers[1], updateCategorySizes: isUpdateVers[2], updateShippings: isUpdateVers[3], updateProductConditions: isUpdateVers[4], updateProvincesRegions: isUpdateVers[5])
+                        } else {
+                            println("Same metadata version")
+                        }
+                        
+                        CDVersion.saveVersions(data)
                     }
                 }
         }
     }
     
-    func updateMetadata()
+    func updateMetadata(updateBrands : String, updateCategories : String, updateCategorySizes : String, updateShippings : String, updateProductConditions : String, updateProvincesRegions : String)
     {
-        println("Updating metadata")
-        request(APIApp.Metadata).responseJSON
+        request(APIApp.Metadata(brands: updateBrands, categories: updateCategories, categorySizes: updateCategorySizes, shippings: updateShippings, productConditions: updateProductConditions, provincesRegions: updateProvincesRegions)).responseJSON
             {_, _, metaRes, metaErr in
                 if (metaErr != nil) { // Terdapat error
                     println("Error getting metadata: \(metaErr!.description)")
@@ -122,14 +155,51 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                     let metaJson = JSON(metaRes!)
                     let metadata = metaJson["_data"]
                     if (metadata == nil) { // Data kembalian kosong
-//                        let obj : [String : String] = res as! [String : String]
-//                        let message = obj["_message"]
-//                        println("Empty metadata, error: \(message)")
+                        println("Error getting metadata")
                     } else { // Berhasil
-                        // Hapus data lama kemudian simpan yang baru // asynchronous!!
+                        // Asynchronous update!!
                         dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), {
-                            if (CDProvince.deleteAll() && CDRegion.deleteAll()) {
-                                CDProvince.saveProvinceRegions(metadata["provinces_regions"])
+                            // Update brands
+                            if (updateBrands == "1") {
+                                println("Updating brands..")
+                                if (CDBrand.deleteAll()) {
+                                    CDBrand.saveBrands(metadata["brands"])
+                                }
+                            }
+                            // Update categories
+                            if (updateCategories == "1") {
+                                println("Updating categories..")
+                                if (CDCategory.deleteAll()) {
+                                    CDCategory.saveCategories(metadata["categories"])
+                                }
+                            }
+                            // Update category sizes
+                            if (updateCategorySizes == "1") {
+                                println("Updating category sizes..")
+                                if (CDCategorySize.deleteAll()) {
+                                    CDCategorySize.saveCategorySizes(metadata["category_sizes"])
+                                }
+                            }
+                            // Update shippings
+                            if (updateShippings == "1") {
+                                println("Updating shippings..")
+                                if (CDShipping.deleteAll()) {
+                                    CDShipping.saveShippings(metadata["shippings"])
+                                }
+                            }
+                            // Update product conditions
+                            if (updateProductConditions == "1") {
+                                println("Updating product conditions..")
+                                if (CDProductCondition.deleteAll()) {
+                                    CDProductCondition.saveProductConditions(metadata["product_conditions"])
+                                }
+                            }
+                            // Update provinces regions
+                            if (updateProvincesRegions == "1") {
+                                println("Updating provinces regions..")
+                                if (CDProvince.deleteAll() && CDRegion.deleteAll()) {
+                                    CDProvince.saveProvinceRegions(metadata["provinces_regions"])
+                                }
                             }
                         })
                     }
@@ -193,7 +263,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         let url = self.applicationDocumentsDirectory.URLByAppendingPathComponent("Prelo.sqlite")
         var error: NSError? = nil
         var failureReason = "There was an error creating or loading the application's saved data."
-        if coordinator!.addPersistentStoreWithType(NSSQLiteStoreType, configuration: nil, URL: url, options: nil, error: &error) == nil {
+        
+        let opt = [NSMigratePersistentStoresAutomaticallyOption:true, NSInferMappingModelAutomaticallyOption:true]
+        
+        if coordinator!.addPersistentStoreWithType(NSSQLiteStoreType, configuration: nil, URL: url, options: opt, error: &error) == nil {
             coordinator = nil
             // Report any error we got.
             var dict = [String: AnyObject]()
@@ -263,54 +336,5 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: NSError) {
         println("ERROR : \(error)")
     }
-    
-    // MARK: - Version Check
-    
-//    func versionCheck() {
-//        request(APIApp.Version(appType: "ios")).responseJSON
-//            {_, _, res, err in
-//                if (err != nil) { // Terdapat error
-//                    println("Error getting version: \(err!.description)")
-//                } else {
-//                    let json = JSON(res!)
-//                    let data = json["_data"]
-//                    if (data == nil) { // Data kembalian kosong
-//                        let obj : [String : String] = res as! [String : String]
-//                        let message = obj["_message"]
-//                        println("Empty version data, error: \(message)")
-//                    } else { // Berhasil
-//                        println("Version data: \(data)")
-//                        
-//                        // Jika versi metadata baru, load dan save kembali di coredata
-//                        let ver : CDVersion? = CDVersion.getOne()
-//                        if (ver?.metadataVersion == data["metadata_version"].string) {
-//                            println("Same metadata version")
-//                        } else {
-//                            println("Updating metadata")
-//                            request(APIApp.Metadata).responseJSON
-//                                {_, _, metaRes, metaErr in
-//                                    if (metaErr != nil) { // Terdapat error
-//                                        println("Error getting metadata: \(metaErr!.description)")
-//                                    } else {
-//                                        let metaJson = JSON(metaRes!)
-//                                        let metadata = metaJson["_data"]
-//                                        if (metadata == nil) { // Data kembalian kosong
-//                                            let obj : [String : String] = res as! [String : String]
-//                                            let message = obj["_message"]
-//                                            println("Empty metadata, error: \(message)")
-//                                        } else { // Berhasil
-//                                            // Hapus data lama kemudian simpan yang baru
-//                                            if (CDProvince.deleteAll() && CDRegion.deleteAll()) {
-//                                                CDProvince.saveProvinceRegions(metadata["provinces_regions"])
-//                                            }
-//                                        }
-//                                    }
-//                            }
-//                        }
-//                        CDVersion.saveVersion(data)
-//                    }
-//                }
-//        }
-//    }
 }
